% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{matrix_mdp_lp}
\alias{matrix_mdp_lp}
\title{Constructs the linear programming matrix for the MDP}
\usage{
matrix_mdp_lp(mdp, discount)
}
\arguments{
\item{mdp}{A dataframe representation of the MDP. Each row
represents a single transition from one state to another
after taking an action a. The columns are:
idstatefrom, idaction, idstateto, probability, reward}

\item{discount}{Discount factor in [0,1]}
}
\value{
A list with entries A, b, and idstateaction which is a dataframe with
        row_index (1-based), stateid (0-based), actionid (0-based) that
        identifies the state and action for each row of the output
}
\description{
The method can construct the LP matrix for the MDP, which is defined as
follows:
A = [I - gamma P_1; I - gamma P_2; ...] where P_a is the transition
probability for action_a. It also constructs the corresponding vector of rewards
}
