% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simulate_mdp}
\alias{simulate_mdp}
\title{Simulates an MDP}
\usage{
simulate_mdp(mdp, initial_state, policy, horizon, episodes, seed = NULL)
}
\arguments{
\item{mdp}{Definition of the MDP}

\item{initial_state}{The index of the initial state}

\item{policy}{Assumes a randomized policy as the input}

\item{horizon}{How many steps to execute for each episode}

\item{episodes}{Number of episodes to run}

\item{seed}{Random number generator seed (a number)}
}
\value{
A dataframe with the states, actions, and rewards received
}
\description{
Simulates an MDP
}
