% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{worstcase_l1}
\alias{worstcase_l1}
\title{Computes the maximum distribution subject to L1 constraints}
\usage{
worstcase_l1(value, reference_dst, budget)
}
\arguments{
\item{value}{Random variable (objective)}

\item{reference_dst}{Reference distribution of the same size as value}

\item{budget}{Maximum L1 distance from the reference dst}
}
\description{
Computes the maximum distribution subject to L1 constraints
}
